backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=4.5.4
pydantic>=2.0.0
pgvector>=0.2.0
httpx>=0.24.0
python-dotenv>=1.0.0
supabase>=1.0.0
Pillow>=9.5.0
requests>=2.31.0
aiohttp>=3.8.4
imagehash>=4.3.1 